moving getJson call to requestUtil + gearing up to start sending mail [wip]
